commit 18.06.2019 added: - return change - getters, setters moneyCapacity - isMoneyInTheMachine - corrected pictureOfTicket - make private variables and methods

+++ src/main/java/TicketMachine/Tickets.java
@@ -6,7 +6,7 @@ public enum Tickets {
 
     double price;
 
-    Tickets(int price) {
+    Tickets(double price) {
         this.price = price;
     }
